+++ muttlib/dbconn/__init__.py
@@ -15,6 +15,8 @@ from muttlib.dbconn.sqlserver import (
     SqlServerClient,
 )
 from muttlib.dbconn.teradata import TERADATA_DB_TYPE, TeradataClient
+from muttlib.dbconn.trino import TRINO_DB_TYPE, TRINO_DIALECT, TrinoClient
+from muttlib.dbconn.redshift import REDSHIFT_DB_TYPE, REDSHIFT_DIALECT, RedshiftClient
 
 connectors = {
     BIGQUERY_DB_TYPE: BigQueryClient,
@@ -27,6 +29,8 @@ connectors = {
     SQLITE_DB_TYPE: SqliteClient,
     SQLSERVER_DB_TYPE: SqlServerClient,
     TERADATA_DB_TYPE: TeradataClient,
+    TRINO_DB_TYPE: TrinoClient,
+    REDSHIFT_DB_TYPE: RedshiftClient,
 }
 
 
@@ -55,6 +59,8 @@ def parse_connection_string(connstr):
     db_type = {
         POSTGRES_DIALECT: POSTGRES_DB_TYPE,
         SQLSERVER_DIALECT: SQLSERVER_DB_TYPE,
+        TRINO_DIALECT: TRINO_DB_TYPE,
+        REDSHIFT_DIALECT: REDSHIFT_DB_TYPE,
     }.get(dialect, dialect)
     rv = {"db_type": db_type, "dialect": dialect}
     if r.database: